Fix problems with PI masking logic and move to numpy.nan for missings
Previously, we coded missing values as 0 (the logic here was still
broken a little bit).  However, this is incorrect as the values
are not zero, they are Null.  So, move to numpy.nan for the masking
and fix functions to compute summary stats assuming numpy.nans are
present.

Also give type argument for --threshold parameter.

I've checked these results with output from PhyDesign, and
everything is close once again (although their masking is a little
tricky).  I still prefer conducting the masking in Python versus
in PhyDesign.